setup typescript and lint
