+++ .vscode/settings.json
@@ -2,7 +2,6 @@
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": true
   },
-  "eslint.validate": ["javascript"],
   "editor.tabSize": 2,
   "eslint.format.enable": true,
   "vetur.format.enable": false
