+++ server/jest.config.js
@@ -1,14 +1,14 @@
 module.exports = {
-  testEnvironment: "node",
+  testEnvironment: 'node',
   testEnvironmentOptions: {
-    NODE_ENV: "test",
+    NODE_ENV: 'test',
   },
   restoreMocks: true,
   coveragePathIgnorePatterns: [
-    "node_modules",
-    "src/config",
-    "src/app.js",
-    "tests",
+    'node_modules',
+    'src/config',
+    'src/app.js',
+    'tests',
   ],
-  coverageReporters: ["text", "lcov", "clover", "html"],
+  coverageReporters: ['text', 'lcov', 'clover', 'html'],
 };
