+++ server/package.json
@@ -1,11 +1,13 @@
 {
-  "name": "scot",
-  "description": "scot service",
+  "name": "scot-api",
+  "description": "scot api service",
   "version": "0.0.0",
-  "main": "app.js",
+  "main": "dist/index.js",
   "private": true,
   "scripts": {
-    "start": "nodemon --exec babel-node index.js",
+    "start": "node dist/index.js",
+    "watch": "nodemon index.ts",
+    "build": "tsc && cp -r public dist/public",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
     "test": "jest -i"
@@ -29,6 +31,10 @@
     "@babel/node": "^7.12.6",
     "@babel/polyfill": "^7.12.1",
     "@babel/preset-env": "^7.12.7",
+    "@types/express": "^4.17.9",
+    "@types/node": "^14.14.14",
+    "@typescript-eslint/eslint-plugin": "^4.10.0",
+    "@typescript-eslint/parser": "^4.10.0",
     "eslint": "^7.15.0",
     "eslint-config-airbnb-base": "^14.0.0",
     "eslint-config-prettier": "^6.4.0",
@@ -39,6 +45,9 @@
     "jest": "^26.6.3",
     "node-mocks-http": "^1.9.0",
     "nodemon": "^2.0.6",
-    "prettier": "^2.2.1"
+    "prettier": "^2.2.1",
+    "ts-jest": "^26.4.4",
+    "ts-node": "^9.1.1",
+    "typescript": "^4.1.3"
   }
 }
