+++ server/src/routes/index.js
@@ -1,9 +1,9 @@
-import express from "express";
+import express from 'express';
 
 const router = express.Router();
 
-router.get("/", (req, res) => {
-  res.json({ x: 1, y: "a" });
+router.get('/', (req, res) => {
+  res.json({ x: 1, y: 'a' });
 });
 
 export default router;
