+++ server/tsconfig.json
@@ -1,0 +1,26 @@
+{
+  "compilerOptions": {
+    "target": "ES2019",
+    "module": "commonjs",
+    "rootDir": ".",
+    "baseUrl": ".",
+    "outDir": "dist",
+    "esModuleInterop": true,
+    "importHelpers": true,
+    "strict": true,
+    "allowJs": true,
+    "sourceMap": true,
+    "moduleResolution": "node",
+    "strictNullChecks": true,
+    "resolveJsonModule": true,
+  },
+  "include": [
+    "src", "index.ts",
+  ],
+  "exclude": [
+    ".vscode",
+    "dist",
+    "public",
+    "node_modules"
+  ]
+}
