--- web/.eslintcache
@@ -1,1 +1,0 @@
-[{"/Users/kakao/workspace/scot/web/src/index.js":"1","/Users/kakao/workspace/scot/web/src/reportWebVitals.js":"2","/Users/kakao/workspace/scot/web/src/App.js":"3"},{"size":500,"mtime":1608278174828,"results":"4","hashOfConfig":"5"},{"size":362,"mtime":1608278174830,"results":"6","hashOfConfig":"5"},{"size":528,"mtime":1608278174825,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"hopyqg",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/Users/kakao/workspace/scot/web/src/index.js",[],"/Users/kakao/workspace/scot/web/src/reportWebVitals.js",[],"/Users/kakao/workspace/scot/web/src/App.js",[]]
