+++ web/.eslintignore
@@ -1,3 +1,3 @@
 node_modules
 public
-dist
+build
